+++ src/tracing/now.js
@@ -1,9 +1,6 @@
 const loadTime = getNanoSeconds();
 const nodeLoadTime = loadTime - process.uptime() * 1e9;
 
-console.log("loadTime", loadTime);
-console.log("nodeLoadTime", nodeLoadTime);
-
 function getNanoSeconds() {
 	const time = process.hrtime();
 	return time[0] * 1e9 + time[1];
@@ -15,8 +12,5 @@ function now() {
 
 const loadNs = now();
 const loadMs = Date.now();
-
-console.log("loadNs", loadNs);
-console.log("loadMs", loadMs);
 
 module.exports = () => loadMs + now() - loadNs;